+++ server/socket.js
@@ -2,7 +2,7 @@ const http = require('./server');
 const io = require('socket.io')(http);
 
 io.on('connection', (socket) => {
-  socket.on('toggle noise', () => {
-    io.emit('toggle noise');
+  socket.on('toggle noise', (playerId) => {
+    io.emit('toggle noise', playerId);
   });
 });