+++ src/components/App.js
@@ -8,6 +8,7 @@ import { fetchPlayer } from '../features/player/playerSlice';
 import AppModeContainer from '../features/appMode/AppModeContainer';
 import NoiseToggleContainer from '../features/noise/NoiseToggleContainer';
 import AddRemoteContainer from '../features/addRemote/AddRemoteContainer';
+import RemoteContainer from '../features/remote/RemoteContainer';
 
 Modal.setAppElement('#root');
 
@@ -26,9 +27,11 @@ const App = () => {
       <header className="App-header">
         <h1>Quick Noise</h1>
         <NoiseToggleContainer />
-        <br/>
+        <br />
         <AppModeContainer />
+        <br />
         <AddRemoteContainer />
+        <RemoteContainer />
       </header>
     </div>
   );